Add a utility lib that merges/unmerges images.
This go around, we implement the Merge() method which merges a cover
image and secret image. The merge operation is done pixel by pixel. The
4 most significant bits of each 8-bit RGB channel in the secret image
are used to overwrite the 4 least significant bits in the corresponding
channel in the cover image. The input cover and secret images are
unaltered. Instead, a third output image representing the merge is
output.