Always write out the merged image as a PNG.
We want to be sure that our merged image's pixels are written out
without alteration. Using the PNG format, we know that the image data
will be compressed using a lossless algorithm. In contrast, a format
like JPEG would mangle one or more pixels making it difficult to recover
the image later.